+++ lib/interact.dart
@@ -1,8 +1,8 @@
-library interact;
+import 'package:interact2/src/framework/framework.dart' show Context;
 
-import 'package:interact/src/framework/framework.dart' show Context;
-
+export 'src/async_input.dart';
 export 'src/confirm.dart';
+export 'src/framework/framework.dart';
 export 'src/input.dart';
 export 'src/multi_progress.dart';
 export 'src/multi_select.dart';
@@ -13,6 +13,8 @@ export 'src/select.dart';
 export 'src/sort.dart';
 export 'src/spinner.dart';
 export 'src/theme/theme.dart';
+export 'src/utils/prompt.dart';
+export 'src/utils/utils.dart';
 
 /// Resets the Terminal to default values.
 void Function() reset = Context.reset;