+++ lib/src/framework/framework.dart
@@ -1,10 +1,13 @@
-library interact.framework;
+library;
 
+import 'dart:async';
+import 'dart:convert';
 import 'dart:io';
 
 import 'package:dart_console/dart_console.dart';
 import 'package:meta/meta.dart';
 
+part 'async_component.dart';
 part 'component.dart';
 part 'context.dart';
 part 'state.dart';