+++ lib/src/spinner.dart
@@ -1,9 +1,9 @@
-import 'dart:async' show Timer, StreamSubscription;
+import 'dart:async' show StreamSubscription, Timer;
 import 'dart:io' show ProcessSignal;
 
-import 'package:interact/src/framework/framework.dart';
-import 'package:interact/src/theme/theme.dart';
-import 'package:interact/src/utils/utils.dart';
+import 'package:interact2/src/framework/framework.dart';
+import 'package:interact2/src/theme/theme.dart';
+import 'package:interact2/src/utils/utils.dart';
 
 String _prompt(bool x) => '';